Moves SurfaceSector into camera subpackage
