+++ src/pyrite/camera/surface_sector.py
